fix(Session): avoid password confirmation on SSO
SSO backends like SAML and OIDC tried a trick to suppress password

confirmations as they are not possible by design. At least for SAML it was

not reliable when existing user backends where used as user repositories.

Now we are setting a special scope with the token, and also make sure that

the scope is taken over when tokens are regenerated.

Signed-off-by: Arthur Schiwon <blizzz@arthur-schiwon.de>

[skip ci]